Upgrade lint and jest packages versions
