+++ .eslintrc
@@ -1,28 +1,11 @@
 {
-  "parser": "babel-eslint",
-  "extends": [
-    "vtex",
-    "vtex-react",
-    "plugin:import/recommended"
-  ],
-  "plugins": ["import"],
-  "rules": {
-    "import/no-unresolved": [
-      "error",
-      {
-        "commonjs": true,
-        "amd": true,
-        "ignore": ["^([a-zA-Z@]+[-\\.]?)+"]
-      }
-    ]
-  },
+  "extends": "vtex",
+  "root": true,
+  "parser": "@typescript-eslint/parser",
   "env": {
-    "browser": true,
-    "node": true,
-    "mocha": true,
-    "jest": true
+    "node": true
   },
-  "globals": {
-    "__DEV__": true
+  "rules": {
+    "react/jsx-filename-extension": "off"
   }
 }
