+++ jest.transform.js
@@ -1,3 +1,11 @@
-module.exports = require('babel-jest').createTransformer({
-  presets: ['react-app'],
+const babelJest = require('babel-jest')
+
+module.exports = babelJest.default.createTransformer({
+  presets: [
+    require.resolve('@babel/preset-env'),
+    require.resolve('@babel/preset-react'),
+    require.resolve('@babel/preset-typescript'),
+  ],
+  babelrc: false,
+  configFile: false,
 })
