--- react/components/__mocks__/vtex.shipping-estimate-translator/TranslateEstimate.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react'
-
-class TranslateEstimate extends Component {
-  render() {
-    return 'TranslateEstimate'
-  }
-}
-
-export default TranslateEstimate
