+++ react/package.json
@@ -7,19 +7,9 @@
     "lib"
   ],
   "scripts": {
-    "build": "nwb build-react-component --copy-files --no-demo && nwb build-react-app && npm run css && npm run removeMocks && npm run addLocales",
     "css": "mv dist/styles.css lib/styles.css && mv dist/styles.css.map lib/styles.css.map ",
     "removeMocks": "rm -rf lib/**/__mocks__ && rm -rf lib/__mocks__ && rm -rf dist",
     "addLocales": "cp -R -f react/locales/ lib/locales/",
-    "clean": "nwb clean-module && nwb clean-demo",
-    "start": "nwb serve-react-demo",
-    "lint": "eslint react/**/*.js",
-    "lint-fix": "eslint --fix react/**/*.js",
-    "test": "jest --env=jsdom",
-    "test:coverage": "jest --env=jsdom --coverage",
-    "test:watch": "jest --env=jsdom --watch",
-    "prepublishOnly": "npm run build",
-    "build:link": "npm link && watch 'npm run build' src",
     "postreleasy": "npm publish --access public"
   },
   "keywords": [
@@ -39,39 +29,11 @@
     "@vtex/estimate-calculator": "^1.1.0",
     "axios": "^0.19.0",
     "axios-retry": "^3.1.2",
-    "babel-plugin-add-react-displayname": "^0.0.5",
     "classnames": "2.2.5",
     "lodash": "^4.17.5",
     "react-infinite-scroller": "^1.2.4",
     "react-intl": "^2.8.0",
-    "react-tooltip": "^3.10.0",
-    "watch": "^1.0.2"
-  },
-  "jest": {
-    "setupTestFrameworkScriptFile": "<rootDir>/react/setupTests.js",
-    "testPathIgnorePatterns": [
-      "/node_modules/",
-      "<rootDir>/es/",
-      "<rootDir>/lib/",
-      "<rootDir>/umd/"
-    ],
-    "modulePathIgnorePatterns": [
-      "<rootDir>/es/",
-      "<rootDir>/lib/",
-      "<rootDir>/umd/"
-    ],
-    "transformIgnorePatterns": [
-      "[/\\\\]lib[/\\\\].+\\.(js|jsx)$",
-      "[/\\\\]umd[/\\\\].+\\.(js|jsx)$",
-      "[/\\\\]es[/\\\\].+\\.(js|jsx)$",
-      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx)$"
-    ],
-    "moduleNameMapper": {
-      "\\.(css|jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)(\\?.*)?$": "identity-obj-proxy"
-    },
-    "transform": {
-      "^.+\\.js$": "<rootDir>/jest.transform.js"
-    }
+    "react-tooltip": "^3.10.0"
   },
   "husky": {
     "hooks": {
@@ -86,12 +48,15 @@
     "lib": "lib"
   },
   "devDependencies": {
+    "@types/node": "^16.11.6",
+    "@types/prop-types": "^15.7.4",
+    "@types/react": "^17.0.34",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-intl": "2",
     "@vtex/address-form": "^3.5.8",
-    "eslint": "^5.15.1",
-    "eslint-config-vtex": "^10.1.0",
-    "eslint-utils": "^1.3.1",
     "vtex.country-codes": "http://vtex.vteximg.com.br/_v/public/typings/v1/vtex.country-codes@1.1.2/public/_types/react",
     "vtex.render-runtime": "http://vtex.vteximg.com.br/_v/public/typings/v1/vtex.render-runtime@7.39.1/public/_types/react",
-    "vtex.shipping-estimate-translator": "http://vtex.vteximg.com.br/_v/public/typings/v1/vtex.shipping-estimate-translator@1.0.4/public/_types/react"
+    "vtex.shipping-estimate-translator": "http://vtex.vteximg.com.br/_v/public/typings/v1/vtex.shipping-estimate-translator@1.0.4/public/_types/react",
+    "watch": "^1.0.2"
   }
 }
