+++ setupTestsAfterEnv.js
@@ -4,6 +4,6 @@ import Adapter from 'enzyme-adapter-react-16'
 
 Enzyme.configure({ adapter: new Adapter() })
 
-global.console.warn = warn => {
+global.console.warn = (warn) => {
   throw new Error(warn)
 }
